[ALPS06512719] disp: unintentional integer overflow
[Detail]
Fix coverity issue. Unintentional integer overflow.

MTK-Commit-Id: bbc3feb8901773a50f831e219a27defa4818a940

CR-Id: ALPS06512719
Feature: Display DSI

+++ drivers/misc/mediatek/video/mt6739/dispsys/ddp_wdma_ex.c
@@ -418,7 +418,7 @@ static int wdma_golden_setting(enum DISP_MODULE_ENUM module,
 
 	DISP_REG_SET(cmdq, offset + DISP_REG_WDMA_BUF_CON4, regval);
 
-	consume_rate = (unsigned long long)(res * frame_rate);
+	consume_rate = ((unsigned long long)res) * ((unsigned long long)frame_rate);
 	do_div(consume_rate, 1000);
 	consume_rate *= 1250;
 	do_div(consume_rate, 16 * 1000);
